refactor: make it easier to add/change behaviour of review process
Created a general interface Reviewer in preparation for adding a
FilteringReviewer that respects .gitignore.
